refactor: create @pnpm/workspace.filter-packages-from-dir
