+++ workspace/filter-packages-from-dir/README.md
@@ -1,0 +1,15 @@
+# @pnpm/workspace.filter-packages-from-dir
+
+> Filters packages in a directory
+
+[![npm version](https://img.shields.io/npm/v/@pnpm/workspace.filter-packages-from-dir.svg)](https://www.npmjs.com/package/@pnpm/workspace.filter-packages-from-dir)
+
+## Installation
+
+```sh
+pnpm add @pnpm/workspace.filter-packages-from-dir
+```
+
+## License
+
+MIT
